requirements.txt
langchain
langchain-community
langchain-core
langchain-ollama
chromadb
pydantic
python-dotenv